7_Agent_Architecture/7.3-MultiAgentIntro/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
boto3==1.39.0
botocore==1.39.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fasta2a==0.3.5
fastapi==0.115.14
fastavro==1.11.1
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.23.0
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.67
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
logfire-api==3.21.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.10.1
mdurl==0.1.2
mistralai==1.9.1
narwhals==1.47.0
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
opentelemetry-api==1.34.1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.3.5
pydantic-ai-slim==0.3.5
pydantic-evals==0.3.5
pydantic-graph==0.3.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.46.1
tenacity==8.5.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
zipp==3.23.0
zstandard==0.23.0
